% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_data.R
\name{get_api_data}
\alias{get_api_data}
\title{Get realtime data from City Bike API}
\usage{

}
\arguments{
\item{client_id}{A character string that is the client identification to access the API.}

\item{data}{A character string that informs the function that you want to retrieve
\code{"availability"} or \code{"stations"} data.}

\item{returnOriginal}{Optional. A logical argument that specifies if you want to retrieve
the original \code{GET} request instead of a dataframe.
Defaults to \code{FALSE}.}
}
\value{
Returns a dataframe if \code{returnOriginal = FALSE}
and a list if \code{returnOriginal = TRUE}
}
\description{
\code{get_api_data} gets realtime data from the
City Bike API for Oslo, Bergen or Trondheim in Norway.
}
